+++ src/high-level/tests/KinDynComputationsUnitTest.cpp
@@ -118,12 +118,13 @@ void testRelativeTransform(iDynTree::KinDynComputations & dynComp)
 void testAverageVelocityAndTotalMomentumJacobian(iDynTree::KinDynComputations & dynComp)
 {
     iDynTree::Twist avgVel;
-    iDynTree::SpatialMomentum mom;
-    iDynTree::Vector6 avgVelCheck, momCheck;
+    iDynTree::SpatialMomentum mom, centroidalMom;
+    iDynTree::Vector6 avgVelCheck, momCheck, centroidalMomCheck;
     iDynTree::VectorDynSize nu(dynComp.getNrOfDegreesOfFreedom()+6);
     dynComp.getModelVel(nu);
 
     MomentumFreeFloatingJacobian momJac(dynComp.getRobotModel());
+    MomentumFreeFloatingJacobian centroidalMomJac(dynComp.getRobotModel());
     FrameFreeFloatingJacobian    avgVelJac(dynComp.getRobotModel());
 
     avgVel = dynComp.getAverageVelocity();
@@ -136,10 +137,16 @@ void testAverageVelocityAndTotalMomentumJacobian(iDynTree::KinDynComputations & 
 
     ASSERT_IS_TRUE(ok);
 
+    centroidalMom = dynComp.getCentroidalTotalMomentum();
+    ok = dynComp.getCentroidalTotalMomentumJacobian(centroidalMomJac);
+    ASSERT_IS_TRUE(ok);
+
     toEigen(momCheck) = toEigen(momJac)*toEigen(nu);
+    toEigen(centroidalMomCheck) = toEigen(centroidalMomJac)*toEigen(nu);
     toEigen(avgVelCheck) = toEigen(avgVelJac)*toEigen(nu);
 
     ASSERT_EQUAL_VECTOR(momCheck,mom.asVector());
+    ASSERT_EQUAL_VECTOR(centroidalMomCheck,centroidalMom.asVector());
     ASSERT_EQUAL_VECTOR(avgVelCheck,avgVel.asVector());
 }